Clean: Relocate STEX MinMaxAccumulator
