+++ src/foundation/MinMaxAccumulator.h
